Add Second Recipe and Index Link, Remove LICENSE
Since this page uses photos from the internet, I didn't feel that
it was appropriate to license this project without better understanding
the implications of doing so without permission from the original
photographers. As such, I removed the license.

The second of three recipe pages were added as part of the project
requirements, along with a link on the index.html page. This one is
a personal recipe for french onion soup.