Add Third Recipe Page, Link, and Remove Template
This commit adds the final recipe page for the project, links to it
from index.html, and removes the template page from the site directory.
The first two changes are made in order to comply with project
requirements, while the last change is made because the template is now
redundant; we aren't iterating any more recipe pages.